# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mafia World Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bodyguards/npc/{guardId}/hire:
    post:
      operationId: hireNpcBodyguard
      tags: [bodyguards]
      summary: Hire an NPC bodyguard
      parameters:
        - in: path
          name: guardId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hired guard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /bodyguards/request:
    post:
      operationId: sendBodyguardRequest
      tags: [bodyguards]
      summary: Send a bodyguard request to another player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BodyguardRequestInput"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyguardRequest"
  /bodyguards/requests/{requestId}/respond:
    post:
      operationId: respondToBodyguardRequest
      tags: [bodyguards]
      summary: Accept or reject a bodyguard request
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BodyguardResponse"
      responses:
        "200":
          description: Response processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /bodyguards/{bodyguardId}/dismiss:
    delete:
      operationId: dismissBodyguard
      tags: [bodyguards]
      summary: Dismiss a bodyguard
      parameters:
        - in: path
          name: bodyguardId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- ATTACKS & SPY ---
...
  /cities/travel:
    post:
      operationId: travelToCity
      tags: [cities]
      summary: Travel to another city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelInput"
      responses:
        "200":
          description: Travel initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelResult"

  # --- ADMIN ---
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    BodyguardRequestInput:
      type: object
      properties:
        targetPlayerId:
          type: integer
        offeredMoney:
          type: integer
          minimum: 0
      required:
        - targetPlayerId
        - offeredMoney
    BodyguardRequest:
      type: object
      properties:
        id:
          type: integer
        fromPlayerId:
          type: integer
        fromUsername:
          type: string
        toPlayerId:
          type: integer
        toUsername:
          type: string
        offeredMoney:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]
        createdAt:
          type: string
      required:
        - id
        - fromPlayerId
        - fromUsername
        - toPlayerId
        - toUsername
        - offeredMoney
        - status
        - createdAt
...
    BodyguardResponse:
      type: object
      properties:
        accept:
          type: boolean
      required:
        - accept
...
    TravelInput:
      type: object
      properties:
        targetCityId:
          type: integer
      required:
        - targetCityId
    TravelResult:
      type: object
      properties:
        travelTimeHours:
          type: number
        arrivalAt:
          type: string
        fromCityName:
          type: string
        toCityName:
          type: string
        message:
          type: string
      required:
        - travelTimeHours
        - arrivalAt
        - fromCityName
        - toCityName
        - message
...